Remove template ui classes, add new ones and restructure dirs
We remove the PageForm classes that were generated as part of the
project template and add new classes. Also the directory stucture for
the UI is changed to have 2 new folders. The Forms folder for all the
menus in the app, and the Components folder for the small components
found in the menus.