Make ImageDatabase check what OS is being used when loading cover image
When loading image files to the file system the application treats
directories as if they are Windows directories. On Linux, loading images
fails because there is a '/' missing at the start of the file path. Now
it will detect the kernel type and add the '/' if it is not a Windows NT
device.